requirements.txt
asgiref==3.2.7
astroid==2.4.1
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
ConfigArgParse==1.2.3
django-filter==2.3.0
et-xmlfile==1.0.1
Flask-BasicAuth==0.2.0
gevent==20.5.1
geventhttpclient==1.4.2
greenlet==0.4.15
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
lazy-object-proxy==1.4.3
locust==1.0.2
lxml==4.5.0
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
MouseInfo==0.1.3
msgpack==1.0.0
openpyxl==3.0.3
Pillow==7.1.2
psutil==5.7.0
pycparser==2.20
PyGetWindow==0.0.8
pylint==2.5.2
PyMsgBox==1.0.8
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.3.4
PyTweening==1.0.3
pytz==2019.3
pyzmq==19.0.1
requests==2.23.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
toml==0.10.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zope.event==4.4
zope.interface==5.1.0
